+++ packages/protocol/test/layer1/preconf/router/PreconfRouterTestBase.sol
@@ -22,7 +22,7 @@ abstract contract PreconfRouterTestBase is Layer1Test {
 
         address taikoWrapper = deploy({
             name: "taiko_wrapper",
-            impl: address(new MockTaikoInbox()),
+            impl: address(new MockTaikoInbox(0)),
             data: abi.encodeCall(MockTaikoInbox.init, (address(0)))
         });
 
@@ -48,7 +48,12 @@ abstract contract PreconfRouterTestBase is Layer1Test {
     function addOperators(address[] memory operators) internal {
         for (uint256 i = 0; i < operators.length; i++) {
             vm.prank(whitelistOwner);
-            whitelist.addOperator(operators[i]);
+            whitelist.addOperator(operators[i], _getSequencerAddress(operators[i]));
         }
+    }
+
+    // Helper function that returns a deterministic sequencer address for testing purposes
+    function _getSequencerAddress(address sequencer) internal pure returns (address) {
+        return address(uint160(sequencer) + 1000);
     }
 }