+++ packages/taiko-client/bindings/metadata/metadata_pacaya.go
@@ -55,7 +55,7 @@ func (m *TaikoDataBlockMetadataPacaya) GetTxListHash() common.Hash {
 	return m.TxsHash
 }
 
-// GetTxListHash returns block extradata.
+// GetExtraData returns block extradata.
 func (m *TaikoDataBlockMetadataPacaya) GetExtraData() []byte {
 	return m.ExtraData[:]
 }
@@ -65,7 +65,7 @@ func (m *TaikoDataBlockMetadataPacaya) GetCoinbase() common.Address {
 	return m.Coinbase
 }
 
-// GetTxListHash returns batch ID.
+// GetBatchID returns batch ID.
 func (m *TaikoDataBlockMetadataPacaya) GetBatchID() *big.Int {
 	return new(big.Int).SetUint64(m.BatchId)
 }