+++ packages/taiko-client/driver/driver_test.go
@@ -13,7 +13,6 @@ import (
 	"github.com/cenkalti/backoff"
 	"github.com/ethereum-optimism/optimism/op-service/eth"
 	"github.com/ethereum-optimism/optimism/op-service/txmgr"
-	"github.com/ethereum/go-ethereum"
 	"github.com/ethereum/go-ethereum/accounts/abi"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/common/hexutil"
@@ -36,6 +35,7 @@ import (
 	preconfblocks "github.com/taikoxyz/taiko-mono/packages/taiko-client/driver/preconf_blocks"
 	"github.com/taikoxyz/taiko-mono/packages/taiko-client/internal/testutils"
 	"github.com/taikoxyz/taiko-mono/packages/taiko-client/pkg/jwt"
+	"github.com/taikoxyz/taiko-mono/packages/taiko-client/pkg/preconf"
 	"github.com/taikoxyz/taiko-mono/packages/taiko-client/pkg/rpc"
 	"github.com/taikoxyz/taiko-mono/packages/taiko-client/pkg/utils"
 	"github.com/taikoxyz/taiko-mono/packages/taiko-client/proposer"
@@ -325,7 +325,9 @@ func (s *DriverTestSuite) TestForcedInclusion() {
 		common.Big0,
 		[]byte{},
 	)
-	s.Nil(err)
+	if err != nil {
+		s.Equal("replacement transaction underpriced", err.Error())
+	}
 	b, err := utils.EncodeAndCompressTxList([]*types.Transaction{forcedInclusionTx})
 	s.Nil(err)
 	s.NotEmpty(b)
@@ -592,9 +594,6 @@ func (s *DriverTestSuite) TestInsertPreconfBlocksWithReorg() {
 		s.Equal(head.Hash(), l1Origin.L2BlockHash)
 		s.Equal(common.Hash{}, l1Origin.L1BlockHash)
 		s.True(l1Origin.IsPreconfBlock())
-
-		_, err = s.RPCClient.L2.HeadL1Origin(context.Background())
-		s.Error(err, ethereum.NotFound)
 	}
 
 	// Propose three same L2 blocks in a batch
@@ -667,6 +666,10 @@ func (s *DriverTestSuite) TestOnUnsafeL2PayloadWithInvalidPayload() {
 }
 
 func (s *DriverTestSuite) TestGossipMessagesRandomReorgs() {
+	if os.Getenv("L2_NODE") != "l2_geth" {
+		s.T().Skip("This test is only applicable for L2 Geth node, since it returns blocks in forks when " +
+			"querying by hash.")
+	}
 	s.ProposeAndInsertEmptyBlocks(s.p, s.d.ChainSyncer().EventSyncer())
 
 	l1Head, err := s.d.rpc.L1.HeaderByNumber(context.Background(), nil)
@@ -704,7 +707,7 @@ func (s *DriverTestSuite) TestGossipMessagesRandomReorgs() {
 
 	s.RevertL1Snapshot(snapshotID)
 	s.L1Mine()
-	s.Nil(rpc.SetHead(context.Background(), s.RPCClient.L2, l2Head1.Number))
+	s.SetHead(l2Head1.Number)
 	_, err = s.RPCClient.L2Engine.SetHeadL1Origin(context.Background(), headL1Origin.BlockID)
 	s.Nil(err)
 	s.d.state.SetL1Current(l1Head)
@@ -730,7 +733,7 @@ func (s *DriverTestSuite) TestGossipMessagesRandomReorgs() {
 
 	s.RevertL1Snapshot(snapshotID)
 	s.L1Mine()
-	s.Nil(rpc.SetHead(context.Background(), s.RPCClient.L2, l2Head1.Number))
+	s.SetHead(l2Head1.Number)
 	_, err = s.RPCClient.L2Engine.SetHeadL1Origin(context.Background(), headL1Origin.BlockID)
 	s.Nil(err)
 	s.d.state.SetL1Current(l1Head)
@@ -806,10 +809,12 @@ func (s *DriverTestSuite) TestGossipMessagesRandomReorgs() {
 
 	ok, err := s.d.ChainSyncer().EventSyncer().BlocksInserterPacaya().IsBasedOnCanonicalChain(
 		context.Background(),
-		&eth.ExecutionPayload{
-			BlockNumber: eth.Uint64Quantity(forkB[len(forkB)-1].Number().Uint64()),
-			BlockHash:   forkB[len(forkB)-1].Hash(),
-			ParentHash:  forkB[len(forkB)-1].ParentHash(),
+		&preconf.Envelope{
+			Payload: &eth.ExecutionPayload{
+				BlockNumber: eth.Uint64Quantity(forkB[len(forkB)-1].Number().Uint64()),
+				BlockHash:   forkB[len(forkB)-1].Hash(),
+				ParentHash:  forkB[len(forkB)-1].ParentHash(),
+			},
 		},
 		headL1Origin,
 	)
@@ -818,10 +823,12 @@ func (s *DriverTestSuite) TestGossipMessagesRandomReorgs() {
 
 	ok, err = s.d.ChainSyncer().EventSyncer().BlocksInserterPacaya().IsBasedOnCanonicalChain(
 		context.Background(),
-		&eth.ExecutionPayload{
-			BlockNumber: eth.Uint64Quantity(forkA[len(forkA)-1].Number().Uint64()),
-			BlockHash:   forkA[len(forkA)-1].Hash(),
-			ParentHash:  forkA[len(forkA)-1].ParentHash(),
+		&preconf.Envelope{
+			Payload: &eth.ExecutionPayload{
+				BlockNumber: eth.Uint64Quantity(forkA[len(forkA)-1].Number().Uint64()),
+				BlockHash:   forkA[len(forkA)-1].Hash(),
+				ParentHash:  forkA[len(forkA)-1].ParentHash(),
+			},
 		},
 		headL1Origin,
 	)
@@ -831,10 +838,12 @@ func (s *DriverTestSuite) TestGossipMessagesRandomReorgs() {
 	if isInForkA {
 		ok, err = s.d.ChainSyncer().EventSyncer().BlocksInserterPacaya().IsBasedOnCanonicalChain(
 			context.Background(),
-			&eth.ExecutionPayload{
-				BlockNumber: eth.Uint64Quantity(forkB[len(forkB)-1].Number().Uint64()),
-				BlockHash:   forkB[len(forkB)-1].Hash(),
-				ParentHash:  forkB[len(forkB)-1].ParentHash(),
+			&preconf.Envelope{
+				Payload: &eth.ExecutionPayload{
+					BlockNumber: eth.Uint64Quantity(forkB[len(forkB)-1].Number().Uint64()),
+					BlockHash:   forkB[len(forkB)-1].Hash(),
+					ParentHash:  forkB[len(forkB)-1].ParentHash(),
+				},
 			},
 			&rawdb.L1Origin{BlockID: headL1Origin.BlockID, L2BlockHash: testutils.RandomHash()},
 		)
@@ -843,10 +852,12 @@ func (s *DriverTestSuite) TestGossipMessagesRandomReorgs() {
 	} else {
 		ok, err = s.d.ChainSyncer().EventSyncer().BlocksInserterPacaya().IsBasedOnCanonicalChain(
 			context.Background(),
-			&eth.ExecutionPayload{
-				BlockNumber: eth.Uint64Quantity(forkA[len(forkA)-1].Number().Uint64()),
-				BlockHash:   forkA[len(forkA)-1].Hash(),
-				ParentHash:  forkA[len(forkA)-1].ParentHash(),
+			&preconf.Envelope{
+				Payload: &eth.ExecutionPayload{
+					BlockNumber: eth.Uint64Quantity(forkA[len(forkA)-1].Number().Uint64()),
+					BlockHash:   forkA[len(forkA)-1].Hash(),
+					ParentHash:  forkA[len(forkA)-1].ParentHash(),
+				},
 			},
 			&rawdb.L1Origin{BlockID: headL1Origin.BlockID, L2BlockHash: testutils.RandomHash()},
 		)
@@ -885,7 +896,7 @@ func (s *DriverTestSuite) TestOnUnsafeL2PayloadWithMissingAncients() {
 	s.Equal(l2Head2.Number.Uint64()-l2Head1.Number().Uint64(), uint64(len(blocks)))
 
 	s.RevertL1Snapshot(snapshotID)
-	s.Nil(rpc.SetHead(context.Background(), s.RPCClient.L2, l2Head1.Number()))
+	s.SetHead(l2Head1.Number())
 	_, err = s.RPCClient.L2Engine.SetHeadL1Origin(context.Background(), headL1Origin.BlockID)
 	s.Nil(err)
 
@@ -904,18 +915,20 @@ func (s *DriverTestSuite) TestOnUnsafeL2PayloadWithMissingAncients() {
 	s.Nil(err)
 	s.GreaterOrEqual(len(l2Head1.Transactions()), 1)
 
-	s.d.preconfBlockServer.PutPayloadsCache(l2Head1.Number().Uint64(), &eth.ExecutionPayload{
-		BlockHash:     l2Head1.Hash(),
-		ParentHash:    l2Head1.ParentHash(),
-		FeeRecipient:  l2Head1.Coinbase(),
-		PrevRandao:    eth.Bytes32(l2Head1.MixDigest()),
-		BlockNumber:   eth.Uint64Quantity(l2Head1.Number().Uint64()),
-		GasLimit:      eth.Uint64Quantity(l2Head1.GasLimit()),
-		Timestamp:     eth.Uint64Quantity(l2Head1.Time()),
-		ExtraData:     l2Head1.Extra(),
-		BaseFeePerGas: eth.Uint256Quantity(*baseFee),
-		Transactions:  []eth.Data{b},
-		Withdrawals:   &types.Withdrawals{},
+	s.d.preconfBlockServer.PutPayloadsCache(l2Head1.Number().Uint64(), &preconf.Envelope{
+		Payload: &eth.ExecutionPayload{
+			BlockHash:     l2Head1.Hash(),
+			ParentHash:    l2Head1.ParentHash(),
+			FeeRecipient:  l2Head1.Coinbase(),
+			PrevRandao:    eth.Bytes32(l2Head1.MixDigest()),
+			BlockNumber:   eth.Uint64Quantity(l2Head1.Number().Uint64()),
+			GasLimit:      eth.Uint64Quantity(l2Head1.GasLimit()),
+			Timestamp:     eth.Uint64Quantity(l2Head1.Time()),
+			ExtraData:     l2Head1.Extra(),
+			BaseFeePerGas: eth.Uint256Quantity(*baseFee),
+			Transactions:  []eth.Data{b},
+			Withdrawals:   &types.Withdrawals{},
+		},
 	})
 
 	// Randomly gossip preconfirmation messages with missing ancients
@@ -1025,18 +1038,20 @@ func (s *DriverTestSuite) TestOnUnsafeL2PayloadWithMissingAncients() {
 	s.Nil(err)
 	s.GreaterOrEqual(len(block.Transactions()), 1)
 
-	s.d.preconfBlockServer.PutPayloadsCache(block.Number().Uint64(), &eth.ExecutionPayload{
-		BlockHash:     block.Hash(),
-		ParentHash:    block.ParentHash(),
-		FeeRecipient:  block.Coinbase(),
-		PrevRandao:    eth.Bytes32(block.MixDigest()),
-		BlockNumber:   eth.Uint64Quantity(block.Number().Uint64()),
-		GasLimit:      eth.Uint64Quantity(block.GasLimit()),
-		Timestamp:     eth.Uint64Quantity(block.Time()),
-		ExtraData:     block.Extra(),
-		BaseFeePerGas: eth.Uint256Quantity(*baseFee),
-		Transactions:  []eth.Data{b},
-		Withdrawals:   &types.Withdrawals{},
+	s.d.preconfBlockServer.PutPayloadsCache(block.Number().Uint64(), &preconf.Envelope{
+		Payload: &eth.ExecutionPayload{
+			BlockHash:     block.Hash(),
+			ParentHash:    block.ParentHash(),
+			FeeRecipient:  block.Coinbase(),
+			PrevRandao:    eth.Bytes32(block.MixDigest()),
+			BlockNumber:   eth.Uint64Quantity(block.Number().Uint64()),
+			GasLimit:      eth.Uint64Quantity(block.GasLimit()),
+			Timestamp:     eth.Uint64Quantity(block.Time()),
+			ExtraData:     block.Extra(),
+			BaseFeePerGas: eth.Uint256Quantity(*baseFee),
+			Transactions:  []eth.Data{b},
+			Withdrawals:   &types.Withdrawals{},
+		},
 	})
 
 	block = getBlock(l2Head1.Number().Uint64() + 2)
@@ -1082,23 +1097,25 @@ func (s *DriverTestSuite) TestSyncerImportPendingBlocksFromCache() {
 		s.Nil(err)
 		s.GreaterOrEqual(len(block.Transactions()), 1)
 
-		s.d.preconfBlockServer.PutPayloadsCache(block.Number().Uint64(), &eth.ExecutionPayload{
-			BlockHash:     block.Hash(),
-			ParentHash:    block.ParentHash(),
-			FeeRecipient:  block.Coinbase(),
-			PrevRandao:    eth.Bytes32(block.MixDigest()),
-			BlockNumber:   eth.Uint64Quantity(block.Number().Uint64()),
-			GasLimit:      eth.Uint64Quantity(block.GasLimit()),
-			Timestamp:     eth.Uint64Quantity(block.Time()),
-			ExtraData:     block.Extra(),
-			BaseFeePerGas: eth.Uint256Quantity(*baseFee),
-			Transactions:  []eth.Data{b},
-			Withdrawals:   &types.Withdrawals{},
+		s.d.preconfBlockServer.PutPayloadsCache(block.Number().Uint64(), &preconf.Envelope{
+			Payload: &eth.ExecutionPayload{
+				BlockHash:     block.Hash(),
+				ParentHash:    block.ParentHash(),
+				FeeRecipient:  block.Coinbase(),
+				PrevRandao:    eth.Bytes32(block.MixDigest()),
+				BlockNumber:   eth.Uint64Quantity(block.Number().Uint64()),
+				GasLimit:      eth.Uint64Quantity(block.GasLimit()),
+				Timestamp:     eth.Uint64Quantity(block.Time()),
+				ExtraData:     block.Extra(),
+				BaseFeePerGas: eth.Uint256Quantity(*baseFee),
+				Transactions:  []eth.Data{b},
+				Withdrawals:   &types.Withdrawals{},
+			},
 		})
 	}
 
 	s.RevertL1Snapshot(snapshotID)
-	s.Nil(rpc.SetHead(context.Background(), s.RPCClient.L2, l2Head1.Number()))
+	s.SetHead(l2Head1.Number())
 	_, err = s.RPCClient.L2Engine.SetHeadL1Origin(context.Background(), headL1Origin.BlockID)
 	s.Nil(err)