+++ packages/taiko-client/pkg/rpc/methods.go
@@ -1044,7 +1044,17 @@ func (c *Client) GetPreconfWhiteListOperator(opts *bind.CallOpts) (common.Addres
 	opts.Context, cancel = CtxWithTimeoutOrDefault(opts.Context, defaultTimeout)
 	defer cancel()
 
-	return c.PacayaClients.PreconfWhitelist.GetOperatorForCurrentEpoch(opts)
+	proposer, err := c.PacayaClients.PreconfWhitelist.GetOperatorForCurrentEpoch(opts)
+	if err != nil {
+		return common.Address{}, fmt.Errorf("failed to get preconfirmation whitelist operator: %w", err)
+	}
+
+	opInfo, err := c.PacayaClients.PreconfWhitelist.Operators(opts, proposer)
+	if err != nil {
+		return common.Address{}, fmt.Errorf("failed to get preconfirmation whitelist operator info: %w", err)
+	}
+
+	return opInfo.SequencerAddress, nil
 }
 
 // GetNextPreconfWhiteListOperator resolves the next preconfirmation whitelist operator address.
@@ -1060,7 +1070,17 @@ func (c *Client) GetNextPreconfWhiteListOperator(opts *bind.CallOpts) (common.Ad
 	opts.Context, cancel = CtxWithTimeoutOrDefault(opts.Context, defaultTimeout)
 	defer cancel()
 
-	return c.PacayaClients.PreconfWhitelist.GetOperatorForNextEpoch(opts)
+	proposer, err := c.PacayaClients.PreconfWhitelist.GetOperatorForNextEpoch(opts)
+	if err != nil {
+		return common.Address{}, fmt.Errorf("failed to get preconfirmation whitelist operator: %w", err)
+	}
+
+	opInfo, err := c.PacayaClients.PreconfWhitelist.Operators(opts, proposer)
+	if err != nil {
+		return common.Address{}, fmt.Errorf("failed to get preconfirmation whitelist operator info: %w", err)
+	}
+
+	return opInfo.SequencerAddress, nil
 }
 
 func (c *Client) GetForcedInclusion(ctx context.Context, isShasta bool) (