+++ packages/taiko-client/prover/prover.go
@@ -140,7 +140,6 @@ func InitFromConfig(
 		p.cfg.TaikoInboxAddress,
 		p.cfg.ProverSetAddress,
 	)
-
 	if txMgr != nil {
 		p.txmgr = txMgr
 	} else {
@@ -153,7 +152,6 @@ func InitFromConfig(
 			return err
 		}
 	}
-
 	if privateTxMgr != nil {
 		p.privateTxmgr = privateTxMgr
 	} else {
@@ -309,7 +307,7 @@ func (p *Prover) requestProofOp(meta metadata.TaikoProposalMetaData) error {
 	return nil
 }
 
-// submitProofsOp performs a batch proof submission operation.
+// submitProofAggregationOp performs a batch proof submission operation.
 func (p *Prover) submitProofAggregationOp(batchProof *proofProducer.BatchProofs) error {
 	submitter := p.proofSubmitterPacaya
 	if utils.IsNil(submitter) {